// src/payload-types.ts
export interface Media {
  id: number;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "pages".
 */
export interface Page {
  id: number;
  title?: string | null;
  slug: string;
  layout: (
    | {
        heading: string;
        subHeading: {
          root: {
            type: string;
            children: {
              type: string;
              version: number;
              [k: string]: unknown;
            }[];
            direction: ('ltr' | 'rtl') | null;
            format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
            indent: number;
            version: number;
          };
          [k: string]: unknown;
        };
        image: number | Media;
        cta_button: {
          label: string;
          url: string;
        };
        id?: string | null;
        blockName?: string | null;
        blockType: 'hero';
      }
    | {
        project: {
          items: {
            id: string | null;
            title: string;
            des?: string | null;
            img?: (number | null) | Media;
            linkname?: string | null;
            link?: string | null;
          }[];
        };
        id?: string | null;
        blockName?: string | null;
        blockType: 'last-project';
      }
    | {
        grit: {
          items: {
            id: string | null;
            title: string;
            description?: string | null;
            className: string;
            img?: (number | null) | Media;
            imgClassName?: string | null;
            titleClassName?: string | null;
            spareImg?: (number | null) | Media;
          }[];
        };
        id?: string | null;
        blockName?: string | null;
        blockType: 'gridItems';
      }
    | {
        clients?:
          | {
              id: string | null;
              name: string;
              logo?: (number | null) | Media;
              nameLogo?: (number | null) | Media;
            }[]
          | null;
        id?: string | null;
        blockName?: string | null;
        blockType: 'client';
      }
  )[];
  seo?: {
    metaTitle?: string | null;
    metaDescription?: string | null;
    /**
     * Gambar ini akan digunakan saat halaman dibagikan di media sosial.
     */
    ogImage?: (number | null) | Media;
  };
  updatedAt: string;
  createdAt: string;
}